feat: Add a widget model specifically for FR env
The FR env has a limited set of features when compared to the commercial
environment. Because of this, we should be limiting those within code so
that we only return the useful widgets in itless environments.

This commit adds a WidgetMappingFR type that specifies a subset of the
commercial widgets.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>