refactor: update widget layout logic
Adds an optional ITLess param in the widget struct. When set to true it
enable the widget in the itless environment. Without it, the widget will
not be displayed.

This makes it a bit easier to manage which widgets appear in the itless
environment as they will all be off by default. It also depends on the
FRONTEND_ENVIRONMENT var rather than feature flags to keep things even
simpler.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>